refactor(ws4): extract record_sequence SMF block to TrackDispatcher+RecordSequence
Pure move of the SMF-import implementation (handleRecordSequenceSMF + 12 helpers
+ nested RecordSequenceReadbackStatus/RecordSequenceNoteValidationError) into a
same-type extension. TrackDispatcher.swift 1256 -> 629 lines. Zero access
changes (entry already internal; all helpers move together so private stays
valid; nested RecordSequenceRegionReadback enum stays in core). Multiset
line-diff: zero original lines changed; golden static-wire diff = 0.
